Update EventService.createEvent method.
Add logic to validate if the creator is a participant.
Add logic to save the creator selected hour when they are a participant.
Add user exception that is thrown when a participant selected hour for an event is invalid.
Update the method parameter used in the API call to `getAutosuggestionByAddress` from `getFormattedAddress` to `getAddressLine`.
Reorder the code.
Change the return type of the method to EventResponse.java
Adapt EventController to the changes.