Add EventInscriptionService for handling participants inscriptions to events.
Add DTOs to map inscription results and requests.

Add EventInscriptionService.createEventInscription method to create/open a new event inscription for the participant that is making the request.

Add EventInscriptionService.completeEventInscription method to complete an event inscription that was open by the participant that is making the request.

Add EventInscriptionService.getEventInscriptionPaymentLink to retrieves the payment link of an inscription of the participant that is making the request.

Add EventInscriptionAlreadyApprovedException.

Add EventInscriptionStateChangeException.