Update EventService.
Update EventRepository.findAll and EventRepository.countAll queries to add filters to only include the events of a specific event creator, to exclude events in which specific user is participant or enrolled, to include only the events in which a specific participant is participant or enrolled.

Update EventServiceImp.getEvents implementation to avoid to include events where the user that is making the request is participant, event creator or enrolled.

Add EventService.getEventsByCreator to retrieves the events created by a specific participant with pagination and the possibility to filter the events by activity, address, date and hours.

Add EventService.getEventsByParticipant to retrieves the events created by a specific participant in which a specific user is participant or enrolled with pagination and the possibility to filter the events by activity, address, date and hours.

Add EventService.findParticipantSignedEventById to retrieves the details of an event in which the user that is making the request is participant or enrolled.

Add EventServiceImp.fetchEventsWithPagination private method to encapsulate the logic for find and maps requested events used in the EventServiceImp.getEvents, EventServiceImp.getEventsByCreator and EventServiceImp.getEventsByParticipant methods.

Renaming event pagination related DTOs.

Add EventWithInscriptionStatusDto..

Add ParticipantNotInscribedException.