Add ParticipantController.
Add endpoint to create/open an event inscription for the participant that is making the request.

Add endpoint to complete an event inscription that was opened by the participant that is making the request.

Add endpoint to retrieve the payment link of an event inscription of the participant that is making the request.

Add endpoint to retrieve the events in which the participant that is making the request is participant or enrolled.

Add endpoint to retrieve the events in which the participant is event creator.

Add endpoint to retrieve the details of an event in which the participant that is making the request is participant or enrolled.