Update event-related application components to address event states.
Update EventDto and EventResumeDto to map EventEntity.state field.

Update EventServiceImp.createEvent implementation to add the logic set the initial states of the events.

Update EventServiceImp.fetchEventsWithPagination method to add the event state value in the build of EventResumeDto instances.

Update EventInscriptionServiceImp.createEventInscription to deny inscriptions when the event state is not WAITING_FOR_PARTICIPANTS.

Update EventInscriptionServiceImp.completeEventInscription to deny inscriptions when the event state is not WAITING_FOR_PARTICIPANTS. Introduce the logic to define the event's start time and to update its state upon reaching the minimum required participants.